doc/Major Explorer First Version/requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
Flask-CORS==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0 